refactor: move cmdtree to std
