+++ std/utils/cmdtree.go
@@ -1,4 +1,4 @@
-package cmd
+package utils
 
 import (
 	"fmt"
